backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.0
groq==0.13.0
httpx==0.27.2
python-multipart==0.0.9